+++ Source/DotNetWorkQueue.Dashboard.Ui/Models/BulkActionResponse.cs
@@ -1,0 +1,7 @@
+namespace DotNetWorkQueue.Dashboard.Ui.Models
+{
+    public class BulkActionResponse
+    {
+        public long Count { get; set; }
+    }
+}