+++ Source/DotNetWorkQueue.Transport.Memory/Basic/MemoryDashboardInit.cs
@@ -78,6 +78,10 @@ namespace DotNetWorkQueue.Transport.Memory.Basic
                 DashboardRequeueErrorMessageCommandHandler>(LifeStyles.Singleton);
             container.Register<ICommandHandlerWithOutput<DashboardResetStaleMessageCommand, long>,
                 DashboardResetStaleMessageCommandHandler>(LifeStyles.Singleton);
+            container.Register<ICommandHandlerWithOutput<DashboardRequeueAllErrorMessagesCommand, long>,
+                DashboardRequeueAllErrorMessagesCommandHandler>(LifeStyles.Singleton);
+            container.Register<ICommandHandlerWithOutput<DashboardResetAllStaleMessagesCommand, long>,
+                DashboardResetAllStaleMessagesCommandHandler>(LifeStyles.Singleton);
             container.Register<ICommandHandlerWithOutput<DashboardUpdateMessageBodyCommand, long>,
                 DashboardUpdateMessageBodyCommandHandler>(LifeStyles.Singleton);
         }