+++ Source/DotNetWorkQueue.Transport.RelationalDatabase/Basic/CommandPrepareHandler/DashboardRequeueAllErrorMessagesPrepareHandler.cs
@@ -1,0 +1,53 @@
+// ---------------------------------------------------------------------
+//This file is part of DotNetWorkQueue
+//Copyright © 2015-2026 Brian Lehnen
+//
+//This library is free software; you can redistribute it and/or
+//modify it under the terms of the GNU Lesser General Public
+//License as published by the Free Software Foundation; either
+//version 2.1 of the License, or (at your option) any later version.
+//
+//This library is distributed in the hope that it will be useful,
+//but WITHOUT ANY WARRANTY; without even the implied warranty of
+//MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+//Lesser General Public License for more details.
+//
+//You should have received a copy of the GNU Lesser General Public
+//License along with this library; if not, write to the Free Software
+//Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+// ---------------------------------------------------------------------
+using System;
+using System.Data;
+using DotNetWorkQueue.Transport.Shared.Basic.Command;
+using DotNetWorkQueue.Validation;
+
+namespace DotNetWorkQueue.Transport.RelationalDatabase.Basic.CommandPrepareHandler
+{
+    /// <inheritdoc />
+    public class DashboardRequeueAllErrorMessagesPrepareHandler : IPrepareCommandHandler<DashboardRequeueAllErrorMessagesCommand>
+    {
+        private readonly CommandStringCache _commandCache;
+        private readonly Lazy<string> _dynamicColumns;
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="DashboardRequeueAllErrorMessagesPrepareHandler"/> class.
+        /// </summary>
+        /// <param name="commandCache">The command cache.</param>
+        /// <param name="optionsFactory">The transport options factory.</param>
+        public DashboardRequeueAllErrorMessagesPrepareHandler(CommandStringCache commandCache, ITransportOptionsFactory optionsFactory)
+        {
+            Guard.NotNull(() => commandCache, commandCache);
+            Guard.NotNull(() => optionsFactory, optionsFactory);
+            _commandCache = commandCache;
+            _dynamicColumns = new Lazy<string>(() => DashboardDynamicColumnHelper.BuildDynamicColumns(optionsFactory.Create()));
+        }
+
+        /// <inheritdoc />
+        public void Handle(DashboardRequeueAllErrorMessagesCommand command, IDbCommand dbCommand, CommandStringTypes commandType)
+        {
+            dbCommand.CommandText = commandType == CommandStringTypes.DashboardRequeueAllErrors
+                ? string.Format(_commandCache.GetCommand(commandType), _dynamicColumns.Value)
+                : _commandCache.GetCommand(commandType);
+        }
+    }
+}